fix(durablefile): keep the mode allowlist comparable on darwin
Stat_t.Mode is uint16 on darwin and uint32 on linux; the allowlist
comparison only ever compiled on linux because make observer's darwin
cross-compile step sat behind red steps and never ran. Widen before
masking.

+++ internal/durablefile/atomic.go
@@ -87,7 +87,9 @@ func regularSingleLinkModes(stat unix.Stat_t, allowed ...uint32) bool {
 		stat.Uid != uint32(os.Geteuid()) {
 		return false
 	}
-	perm := stat.Mode & 0o777
+	// uint32() is load-bearing: Stat_t.Mode is uint16 on darwin and uint32 on
+	// linux, and the darwin cross-compile in `make observer` must keep building.
+	perm := uint32(stat.Mode) & 0o777
 	for _, mode := range allowed {
 		if perm == mode {
 			return true
